requirements.txt
gevent==1.1.2
pytest==2.9.2
pytest-mock==1.1
requests==2.10.0
